Moved stuff on the correct location
