--- frontend/src/app/components/footer/footer.component.html
@@ -1,4 +1,0 @@
-<footer>
-  <span>Version 1.0.0</span> | <span>Cevi QR-Code-Generator</span> | <span>Hosted and developed by <a href="https://cevi.tools">Cevi.Tools</a></span> |
-  <span>Source auf <a href="https://github.com/cevi/qr-code-generator">Github</a></span>
-</footer>
