--- frontend/src/app/components/header/header.component.html
@@ -1,10 +1,0 @@
-<header class="header-bar">
-  <div class="textwrapper">
-    <h1>Cevi QR-Code-Generator</h1>
-  </div>
-  <div class="logowrapper">
-    <a href="https://qr.cevi.tools/" class="logo">
-      <img alt="Kurzes Logo Cevi Schweiz" class="image -short" src="assets/images/logo.svg" />
-    </a>
-  </div>
-</header>
