+++ frontend/src/app/core/app-state.ts
