+++ frontend/src/app/design-switcher/components/design-switcher/design-switcher.component.ts
