+++ frontend/src/app/design-switcher/design-switcher.module.ts
