+++ frontend/src/app/design/components/footer/footer.component.html
@@ -1,0 +1,4 @@
+<footer>
+  <span>Version {{version}}</span> | <span>Cevi QR-Code-Generator</span> | <span>Hosted and developed by <a href="https://cevi.tools">Cevi.Tools</a></span> |
+  <span>Source auf <a href="https://github.com/cevi/qr-code-generator">Github</a></span>
+</footer>
