+++ frontend/src/app/design/components/footer/footer.component.scss
