+++ frontend/src/app/design/components/footer/footer.component.ts
@@ -1,10 +1,11 @@
 import { Component } from '@angular/core';
+import packageJson from '../../../../../package.json';
 
 @Component({
   selector: 'app-footer',
   templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.scss']
+  styleUrls: ['./footer.component.scss'],
 })
 export class FooterComponent {
-
+  public version: string = packageJson.version;
 }
