+++ frontend/src/app/design/components/header/header.component.html
@@ -1,0 +1,10 @@
+<header class="header-bar">
+  <div class="textwrapper">
+    <h1>Cevi QR-Code-Generator</h1>
+  </div>
+  <div class="logowrapper">
+    <a class="logo" href="https://qr.cevi.tools/">
+      <img alt="Kurzes Logo Cevi Schweiz" class="image -short" src="../../../../assets/images/logo.svg" />
+    </a>
+  </div>
+</header>
