+++ frontend/src/app/design/components/header/header.component.scss
