+++ frontend/src/app/type-switcher/components/link/link.component.html
