+++ frontend/src/app/type-switcher/components/phone/phone.component.ts
