+++ frontend/src/app/type-switcher/components/type-switcher/type-switcher.component.html
