+++ frontend/src/app/type-switcher/components/wifi/wifi.component.html
