+++ frontend/src/app/type-switcher/type-switcher.module.ts
