+++ frontend/src/app/ui/text-input/text-input.component.scss
